requirements.txt
streamlit
face_recognition
opencv-python
numpy
Pillow
